requirements.txt
uvicorn~=0.24.0.post1
fastapi~=0.104.1
redis~=5.0.1
pytest~=7.4.3
pydantic~=2.4.2
python-dotenv~=1.0.0
